api/add show user endpoint

+++ apps/api/src/app/http/controllers/users/ban-user/controller.ts
@@ -1,10 +1,13 @@
-import { Request, Response } from "express";
+import { Response } from "express";
 import "dotenv/config";
 import { deleteUser } from "@app/repositories/user.repository";
+import { BanUserRequest } from "./schema";
 
-export const banUserController = async (req: Request, res: Response) => {
+export const banUserController = async (req: BanUserRequest, res: Response) => {
   try {
     const { userId } = req.params;
+
+    if(!userId) return res.status(422).send({message: "Param userId not provided"})
 
     await deleteUser(userId);
